// apps/web/src/socket/msS2/events/client/requestType.ts
enum RequestTypes {
  ADD_SECTION = "ADD_SECTION",
  ATTACH_PATIENT = "ATTACH_PATIENT",
  CHANGE_SECTION = "CHANGE_SECTION",
  CANCEL_CREATE_NOTE = "CANCEL_CREATE_NOTE",
  CLICK_INSERT_SCRIPT = "CLICK_INSERT_SCRIPT",
  CLICK_NOTES_PULL = "CLICK_NOTES_PULL",
  CLICK_REFRESH_VITALS_ON_SCREEN = "CLICK_REFRESH_VITALS_ON_SCREEN",
  COMPOSITION_SUBMISSION_SUCCESSFUL = "COMPOSITION_SUBMISSION_SUCCESSUFL",
  CREATE_NOTE = "CREATE_NOTE",
  CREATE_NOTE_WITH_NOTETYPE = "CREATE_NOTE_WITH_NOTETYPE",
  CREATE_NOTE_WITH_NOTETYPE_PATIENT = "CREATE_NOTE_WITH_NOTETYPE_PATIENT",
  DELETE_SECTION = "DELETE_SECTION",
  SAVE_COMPOSITION_TO_UNFINISHED = "SAVE_COMPOSITION_TO_UNFINISHED",
  UPDATE_APPOINTMENT = "UPDATE_APPOINTMENT",
  UPDATE_PATIENT = "UPDATE_PATIENT",
  UPDATE_SECTION = "UPDATE_SECTION"
}

export { RequestTypes };